requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.77
botocore==1.29.77
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
Django==3.2.16
django-ckeditor==6.3.2
django-countries==7.5
django-extensions==3.2.1
django-formtools==2.4
django-js-asset==2.0.0
django-mailjet==0.3.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-storages==1.13.2
django-taggit==3.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
jmespath==1.0.1
kombu==5.2.4
mailjet-rest==1.3.4
multidict==6.0.4
openai==0.26.5
outcome==1.2.0
phonenumbers==8.13.5
phonenumberslite==8.13.3
Pillow==9.3.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7
requests==2.28.1
s3transfer==0.6.0
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
zope.interface==5.5.2
